pet body part's position and scale properties type was changed from int to float

+++ Application.Services/DataTransferObjects/Creating/PetCreatingDto.cs
@@ -8,20 +8,20 @@ namespace Application.Services.DataTransferObjects.Creating
         public Guid FarmId { get; set; }
 
         public string BodyPicName { get; set; }
-        public int BodyPictureX { get; set; }
-        public int BodyPictureY { get; set; }
+        public float BodyPictureX { get; set; }
+        public float BodyPictureY { get; set; }
         public float BodyPictureScale { get; set; }
         public string EyesPicName { get; set; }
-        public int EyesPictureX { get; set; }
-        public int EyesPictureY { get; set; }
+        public float EyesPictureX { get; set; }
+        public float EyesPictureY { get; set; }
         public float EyesPictureScale { get; set; }
         public string MouthPicName { get; set; }
-        public int MouthPictureX { get; set; }
-        public int MouthPictureY { get; set; }
+        public float MouthPictureX { get; set; }
+        public float MouthPictureY { get; set; }
         public float MouthPictureScale { get; set; }
         public string NosePicName { get; set; }
-        public int NosePictureX { get; set; }
-        public int NosePictureY { get; set; }
+        public float NosePictureX { get; set; }
+        public float NosePictureY { get; set; }
         public float NosePictureScale { get; set; }
     }
 }
